minor updates: changed order of char calls on login to set skills before calling ship fx added class identifier to gm call log added attribute name to debug list (and removed from todo list) added/changed some notes added more fx logging fixed minor decel error in warp removed online effect from type map

+++ src/eve-server/inventory/Inventory.cpp
@@ -439,7 +439,7 @@ InventoryItemRef Inventory::FindFirstByFlag(EVEItemFlags flag) const {
         return InventoryItemRef(nullptr);
 }
 
-InventoryItemRef Inventory::GetByTypeFlag(uint32 typeID, EVEItemFlags flag) const {
+InventoryItemRef Inventory::GetByTypeFlag(uint16 typeID, EVEItemFlags flag) const {
     auto range = m_contentsByFlag.equal_range(flag);
     for ( auto itr = range.first; itr != range.second; ++itr )
         if (itr->second->typeID() == typeID)
